# back/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nareshka",
    "description": "Nareshka - платформа для изучения программирования",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v2/interviews/": {
      "get": {
        "tags": [
          "interviews"
        ],
        "summary": "Получить список интервью",
        "description": "Возвращает список интервью с поддержкой фильтрации по компании и пагинации",
        "operationId": "get_interviews_api_v2_interviews__get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Номер страницы",
              "default": 1,
              "title": "Page"
            },
            "description": "Номер страницы"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Количество записей на странице",
              "default": 20,
              "title": "Limit"
            },
            "description": "Количество записей на странице"
          },
          {
            "name": "company",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по названию компании (устарел, используйте companies)",
              "title": "Company"
            },
            "description": "Фильтр по названию компании (устарел, используйте companies)"
          },
          {
            "name": "companies",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по списку компаний",
              "title": "Companies"
            },
            "description": "Фильтр по списку компаний"
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Поиск по содержимому интервью",
              "title": "Search"
            },
            "description": "Поиск по содержимому интервью"
          },
          {
            "name": "has_audio",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Фильтр по наличию аудио/видео записи",
              "title": "Has Audio"
            },
            "description": "Фильтр по наличию аудио/видео записи"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewsListResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/interviews/company/{company_name}/stats": {
      "get": {
        "tags": [
          "interviews"
        ],
        "summary": "Статистика по компании",
        "description": "Возвращает подробную статистику по интервью конкретной компании",
        "operationId": "get_company_stats_api_v2_interviews_company__company_name__stats_get",
        "parameters": [
          {
            "name": "company_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Company Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyStatsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v2/interviews/analytics/overview": {
      "get": {
        "tags": [
          "interviews"
        ],
        "summary": "Общая аналитика",
        "description": "Возвращает сводную аналитику по всем интервью в системе",
        "operationId": "get_analytics_overview_api_v2_interviews_analytics_overview_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/v2/interviews/companies/list": {
      "get": {
        "tags": [
          "interviews"
        ],
        "summary": "Список компаний",
        "description": "Возвращает список всех компаний с количеством вопросов для использования в фильтрах",
        "operationId": "get_companies_list_api_v2_interviews_companies_list_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompaniesListWithCountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
    "/api/v2/interviews/detail/{interview_id}": {
      "get": {
        "tags": [
          "interviews"
        ],
        "summary": "Получить детали интервью",
        "description": "Возвращает полную информацию об интервью включая полный текст",
        "operationId": "get_interview_detail_api_v2_interviews_detail__interview_id__get",
        "parameters": [
          {
            "name": "interview_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Interview Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InterviewDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyticsResponse": {
        "properties": {
          "total_interviews": {
            "type": "integer",
            "title": "Total Interviews"
          },
          "total_companies": {
            "type": "integer",
            "title": "Total Companies"
          },
          "top_companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyStatsResponse"
            },
            "type": "array",
            "title": "Top Companies"
          },
          "monthly_stats": {
            "items": {
              "type": "object"
            },
            "type": "array",
            "title": "Monthly Stats"
          }
        },
        "type": "object",
        "required": [
          "total_interviews",
          "total_companies",
          "top_companies",
          "monthly_stats"
        ],
        "title": "AnalyticsResponse",
        "description": "Общая аналитика"
      },
...
      "CompaniesListWithCountResponse": {
        "properties": {
          "companies": {
            "items": {
              "$ref": "#/components/schemas/CompanyWithCountResponse"
            },
            "type": "array",
            "title": "Companies"
          }
        },
        "type": "object",
        "required": [
          "companies"
        ],
        "title": "CompaniesListWithCountResponse",
        "description": "Список компаний с количеством вопросов"
      },
...
      "CompanyStatsResponse": {
        "properties": {
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "total_interviews": {
            "type": "integer",
            "title": "Total Interviews"
          },
          "avg_duration": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avg Duration"
          }
        },
        "type": "object",
        "required": [
          "company_name",
          "total_interviews",
          "avg_duration"
        ],
        "title": "CompanyStatsResponse",
        "description": "Статистика по компании"
      },
      "CompanyWithCountResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "count": {
            "type": "integer",
            "title": "Count"
          }
        },
        "type": "object",
        "required": [
          "name",
          "count"
        ],
        "title": "CompanyWithCountResponse",
        "description": "Компания с количеством вопросов"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "InterviewDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "interview_date": {
            "type": "string",
            "format": "date-time",
            "title": "Interview Date"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "full_content": {
            "type": "string",
            "title": "Full Content"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "companies": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Companies"
          },
          "extracted_urls": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Extracted Urls"
          },
          "has_audio_recording": {
            "type": "boolean",
            "title": "Has Audio Recording"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_name",
          "interview_date",
          "position",
          "full_content",
          "tags",
          "companies",
          "extracted_urls",
          "has_audio_recording"
        ],
        "title": "InterviewDetailResponse",
        "description": "Детальный ответ с полным контентом"
      },
      "InterviewRecordResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "company_name": {
            "type": "string",
            "title": "Company Name"
          },
          "interview_date": {
            "type": "string",
            "format": "date-time",
            "title": "Interview Date"
          },
          "position": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Position"
          },
          "full_content": {
            "type": "string",
            "title": "Full Content"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          },
          "has_audio_recording": {
            "type": "boolean",
            "title": "Has Audio Recording"
          }
        },
        "type": "object",
        "required": [
          "id",
          "company_name",
          "interview_date",
          "position",
          "full_content",
          "tags",
          "has_audio_recording"
        ],
        "title": "InterviewRecordResponse",
        "description": "Ответ с данными интервью"
      },
      "InterviewsListResponse": {
        "properties": {
          "interviews": {
            "items": {
              "$ref": "#/components/schemas/InterviewRecordResponse"
            },
            "type": "array",
            "title": "Interviews"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_prev": {
            "type": "boolean",
            "title": "Has Prev"
          }
        },
        "type": "object",
        "required": [
          "interviews",
          "total",
          "page",
          "limit",
          "has_next",
          "has_prev"
        ],
        "title": "InterviewsListResponse",
        "description": "Список интервью с пагинацией"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}